refacor: rename group-and-page slicer to userSettings and update refernces
